refactor: 불필요한 validation 제거
